requirements.txt
requests
matplotlib
shapely
geopandas
autogen
google-generativeai
python-dotenv
